+++ src/plugins/canvas/components/page-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'pagePicker',
-  REQUEST_COUNT: 5,
 };